requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.6
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
pypdf>=4.2.0
google-search-results>=2.4.2
python-dotenv>=1.0.1
